[CAD-2060] Moving and updating dependencies so we can use it in db-sync.
